refactor: domain 하위 entity 패키지 생성 및 User entity 이동
