+++ src/main/java/com/catcher/resource/UserApiController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.catcher.core.domain.User;
+import com.catcher.core.domain.entity.User;
 
 @RequiredArgsConstructor
 @RestController
